racingcar refactor: 테스트 코드 폴더링
